+++ ui/src/types/dashboards.ts
@@ -115,6 +115,12 @@ export interface DashboardName {
 
 interface DashboardFileMetaSection {
   chronografVersion?: string
+  sources?: {
+    [x: string]: {
+      name: string
+      link: string
+    }
+  }
 }
 
 export interface DashboardFile {